feat: store transaction

+++ database/migrations/2023_10_15_034222_create_transactions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('ticket_id');
             $table->foreignId('user_id');
             $table->tinyInteger('status');// 0 unpaid, 1 Paid, 3 canceled
-            $table->double('total_price', 12, 2);
+            $table->integer('total_price');
             $table->boolean('is_travel_assurance');
             $table->boolean('is_delay_assurance');
         });
